TTA_HL_Week_04_Pandas
Requires the .py file named "TTA_DataScience_Week04_csv creation.py" and the "Destinations.json" file to work

Requires the following libraries:
1) pandas
2) numpy
3) matplotlib
4) json
5) python-pptx

The python file imports the data from the reference json file. It then create and manipulate the dataframe and save the result as a csv file named "csv_summary.csv".
It then creates a few plots from the data frame and save them as jpeg.
It finally reads an internal dictionary and creates a PowerPoint file that presents the work and its analysis.